+++ CHANGES.md
@@ -67,6 +67,8 @@ Changes in 11.0
 Changes in 10.9.8
 -----------------
 - Fix remove "-master" suffix from add-on folder in Modules.inc.php & Plugins.inc.php
+- Update Default School Year to 2023 in config.inc.sample.php, rosariosis.sql & rosariosis_mysql.sql
+
 Changes in 10.9.7
 -----------------
 - Fix SQL for Class Average calculation, exclude NULL grades in Grades.fnc.php